+++ frontend/app/src/main/java/com/cs222/nutrify/ui/LoginFragment.kt
@@ -1,0 +1,20 @@
+package com.cs222.nutrify.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.cs222.nutrify.databinding.FragmentLoginBinding
+
+class LoginFragment : Fragment() {
+    private lateinit var binding: FragmentLoginBinding
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentLoginBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+}
