+++ frontend/app/src/main/java/com/cs222/nutrify/ui/MainActivity.kt
@@ -1,13 +1,13 @@
-package com.cs222.nutrify.ui.nutritioninput
+package com.cs222.nutrify.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.cs222.nutrify.R
 
-class NutritionInputActivity : AppCompatActivity() {
+class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_nutritioninput)
+        setContentView(R.layout.activity_main)
     }
 }
