+++ frontend/app/src/main/res/layout/fragment_nutritioninput.xml
@@ -1,14 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
     <ScrollView
         android:id="@+id/scroll"
-        android:layout_width="409dp"
-        android:layout_height="600dp"
+        android:layout_width="300dp"
+        android:layout_height="500dp"
         android:layout_marginStart="27dp"
         android:layout_marginTop="7dp"
         android:layout_marginEnd="27dp"
@@ -101,6 +100,20 @@
     </ScrollView>
 
     <Button
+        android:id="@+id/addNutritionInfo"
+        android:layout_width="wrap_content"
+        android:layout_height="50dp"
+        android:text="Add"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintHeight_percent=".5"
+        app:layout_constraintHorizontal_bias="0.5"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/scroll"
+        app:layout_constraintVertical_bias="0.25"
+        app:layout_constraintWidth_percent=".5" />
+
+    <Button
         android:id="@+id/submit"
         android:layout_width="wrap_content"
         android:layout_height="50dp"
@@ -109,6 +122,7 @@
         app:layout_constraintHeight_percent=".5"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/scroll"
+        app:layout_constraintTop_toBottomOf="@+id/addNutritionInfo"
         app:layout_constraintWidth_percent=".5"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
